# backend/docs/openapi/openapi.yaml
...
openapi: 3.1.0
paths:
...
  /loan-proposals:
    get:
      summary: Listar todas as propostas
      tags:
      - Loan Proposals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
    post:
      summary: Criar uma nova proposta de crédito
      tags:
      - Loan Proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanProposalRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanSimulationResultDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - admin
        - client
  /loan-proposals/available-margin/{customerId}:
    get:
      summary: Obter margem disponível para o cliente
      tags:
      - Loan Proposals
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: number
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - client
  /loan-proposals/installments-pay/{installmentId}/pay/{accountId}/receipt:
    post:
      summary: Pagar parcela e retornar comprovante em PDF
      tags:
      - Loan Proposals
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: installmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - client
        - admin
  /loan-proposals/installments/pay:
    post:
      summary: Pagar uma parcela específica de uma proposta
      tags:
      - Loan Proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInstallmentRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - client
        - admin
  /loan-proposals/my-proposals:
    get:
      summary: Listar propostas do cliente logado com paginação
      tags:
      - Loan Proposals
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultLoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - client
  /loan-proposals/paged:
    get:
      summary: Listar propostas com paginação e filtro por nome/cpf/email
      tags:
      - Loan Proposals
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: search
        in: query
        schema:
          type: string
          default: ""
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultLoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - client
  /loan-proposals/simulate:
    post:
      summary: Simular uma nova proposta de crédito (sem persistir)
      tags:
      - Loan Proposals
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanProposalRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanSimulationResultDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - admin
        - client
  /loan-proposals/{id}:
    put:
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanProposalRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Update
      security:
      - SecurityScheme:
        - admin
  /loan-proposals/{id}/cancel:
    put:
      summary: Marcar proposta como cancelada
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
  /loan-proposals/{id}/ccb:
    get:
      summary: Obter detalhes da CCB (Contrato de Cessão de Crédito)
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
        - client
  /loan-proposals/{id}/ccb-generate:
    get:
      summary: Gerar CCB (Contrato de Cessão de Crédito) em PDF
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
        - client
  /loan-proposals/{id}/ccb-sign:
    get:
      summary: Assinar CCB (Contrato de Cessão de Crédito)
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
        - client
  /loan-proposals/{id}/pay:
    put:
      summary: Marcar proposta como paga
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
  /loan-proposals/{id}/sign:
    put:
      summary: Confirmar assinatura da proposta
      tags:
      - Loan Proposals
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProposalResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - "company, client"
  /pix-keys:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PixKeyRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Pix Key Controller
  /pix-keys/account/{accountId}:
    get:
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get All By Account
      tags:
      - Pix Key Controller
  /pix-keys/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PixKeyRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Pix Key Controller
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get By Id
      tags:
      - Pix Key Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Delete
      tags:
      - Pix Key Controller
...
components:
  schemas:
...
    LoanInstallmentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        value:
          type: number
        dueDate:
          $ref: "#/components/schemas/LocalDate"
        iof:
          type: number
        balance:
          type: number
        interest:
          type: number
        principal:
          type: number
        additionalValue:
          type: number
        presentValue:
          type: number
        status:
          type: string
        period:
          type: integer
          format: int32
        daysElapsed:
          type: integer
          format: int32
        daysLate:
          type: integer
          format: int32
        paidValue:
          type: number
        paid:
          type: boolean
    LoanProposalRequestDTO:
      type: object
      required:
      - customerId
      - companyId
      - requestedAmount
      - availableLimit
      - termInMonths
      - convenioType
      - monthlyInterestRate
      - firstInstallmentDate
      - modoSimulacao
      properties:
        customerId:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        requestedAmount:
          type: number
          minimum: 0.0
        availableLimit:
          type: number
          minimum: 0.0
        termInMonths:
          type: integer
          format: int32
          maximum: 96
          minimum: 1
        convenioType:
          type: string
          pattern: \S
        monthlyInterestRate:
          type: number
          minimum: 0.0
        firstInstallmentDate:
          $ref: "#/components/schemas/LocalDate"
        modoSimulacao:
          type: string
          pattern: \S
        grossAmount:
          type: number
        iofTotal:
          type: number
        iofAnnualRate:
          type: number
        iofAdditionalRate:
          type: number
        proposalDate:
          $ref: "#/components/schemas/LocalDate"
        tacValue:
          type: number
        spread:
          type: number
        effectiveMonthlyRate:
          type: number
        effectiveAnnualRate:
          type: number
        contractedAnnualRate:
          type: number
        productType:
          type: string
        fund:
          type: string
        installmentType:
          type: string
        disbursementValue:
          type: number
        disbursementDate:
          $ref: "#/components/schemas/LocalDate"
        processingCostTotal:
          type: number
        amortizationType:
          type: string
        interestComposition:
          type: string
        ccb:
          type: string
    LoanProposalResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        customerName:
          type: string
        companyName:
          type: string
        requestedAmount:
          type: number
        availableLimit:
          type: number
        termInMonths:
          type: integer
          format: int32
        convenioType:
          type: string
        status:
          $ref: "#/components/schemas/LoanProposalStatus"
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
        installmentValue:
          type: number
        totalPayment:
          type: number
        financedAmount:
          type: number
        monthlyRate:
          type: number
        firstInstallmentDate:
          type: string
        lastInstallmentDate:
          type: string
        numberOfInstallments:
          type: integer
          format: int32
        grossAmount:
          type: number
        iofTotal:
          type: number
        iofAnnualRate:
          type: number
        iofAdditionalRate:
          type: number
        proposalDate:
          $ref: "#/components/schemas/LocalDate"
        tacValue:
          type: number
        spread:
          type: number
        effectiveMonthlyRate:
          type: number
        effectiveAnnualRate:
          type: number
        contractedAnnualRate:
          type: number
        productType:
          type: string
        fund:
          type: string
        installmentType:
          type: string
        disbursementValue:
          type: number
        disbursementDate:
          $ref: "#/components/schemas/LocalDate"
        processingCostTotal:
          type: number
        amortizationType:
          type: string
        interestComposition:
          type: string
        ccb:
          type: string
        installments:
          type: array
          items:
            $ref: "#/components/schemas/LoanInstallmentDTO"
    LoanProposalStatus:
      type: string
      enum:
      - SIMULATED
      - PENDING_SIGNATURE
      - SIGNED
      - UNDER_REVIEW
      - APPROVED
      - REJECTED
      - PAID
      - CANCELED
    LoanSimulationResultDTO:
      type: object
      properties:
        installmentValue:
          type: number
        totalPayment:
          type: number
        financedAmount:
          type: number
        grossAmount:
          type: number
        iofTotal:
          type: number
        iofAnnualRate:
          type: number
        iofAdditionalRate:
          type: number
        proposalDate:
          $ref: "#/components/schemas/LocalDate"
        firstInstallmentDate:
          $ref: "#/components/schemas/LocalDate"
        lastInstallmentDate:
          $ref: "#/components/schemas/LocalDate"
        numberOfInstallments:
          type: integer
          format: int32
        tacValue:
          type: number
        spread:
          type: number
        monthlyRate:
          type: number
        effectiveMonthlyRate:
          type: number
        effectiveAnnualRate:
          type: number
        contractedAnnualRate:
          type: number
        productType:
          type: string
        fund:
          type: string
        installmentType:
          type: string
        disbursementValue:
          type: number
        disbursementDate:
          $ref: "#/components/schemas/LocalDate"
        processingCostTotal:
          type: number
        amortizationType:
          type: string
        interestComposition:
          type: string
        ccb:
          type: string
        proposal:
          type: string
        installments:
          type: array
          items:
            $ref: "#/components/schemas/LoanInstallmentDTO"
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
...
    PagedResultLoanProposalResponseDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LoanProposalResponseDTO"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    PayInstallmentRequestDTO:
      type: object
      properties:
        proposalId:
          type: integer
          format: int64
        installmentNumber:
          type: integer
          format: int32
        bankAccountId:
          type: integer
          format: int64
    PixKeyRequestDTO:
      type: object
      properties:
        pixKey:
          type: string
        keyType:
          type: string
        bankAccountId:
          type: integer
          format: int64
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8080/realms/crediflow/.well-known/openid-configuration
      description: Authentication
...
info:
  title: backend API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8082
  description: Auto generated value
- url: http://0.0.0.0:8082
  description: Auto generated value